Move `package` subcommands to `cli/package.py`
